+++ src/test/java/ee/sk/smartid/rest/SmartIdRestIntegrationTest.java
@@ -37,17 +37,19 @@ import org.junit.Test;
 import java.util.Collections;
 import java.util.concurrent.TimeUnit;
 
+import static ee.sk.test.smartid.integration.SmartIdIntegrationTest.TEST_AGAINST_SMART_ID_DEMO;
 import static java.util.Arrays.asList;
 import static junit.framework.TestCase.assertEquals;
 import static junit.framework.TestCase.assertNotNull;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
+import static org.junit.Assume.assumeTrue;
 
 public class SmartIdRestIntegrationTest {
 
   private static final String RELYING_PARTY_UUID = "00000000-0000-0000-0000-000000000000";
   private static final String RELYING_PARTY_NAME = "DEMO";
-  private static final String DOCUMENT_NUMBER = "PNOEE-30303039816-MD2B-Q";
+  private static final String DOCUMENT_NUMBER = "PNOLT-30303039903-FMB7-Q";
   private static final String DOCUMENT_NUMBER_LT = "PNOLT-30303039914-PBZK-Q";
   private static final String DATA_TO_SIGN = "Hello World!";
   private static final String CERTIFICATE_LEVEL_QUALIFIED = "QUALIFIED";
@@ -56,6 +58,9 @@ public class SmartIdRestIntegrationTest {
   @Before
   public void setUp() {
     connector = new SmartIdRestConnector("https://sid.demo.sk.ee/smart-id-rp/v2/");
+
+    // this allows to switch off tests going against smart-id demo env
+    assumeTrue(TEST_AGAINST_SMART_ID_DEMO);
   }
 
   @Test