+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [2.1.3] - UPCOMING
 
 ### Fixed
-- Possible NPE fix
-- Changed test document number to make two tests pass 
+- Possible NPE fix (in rare cases under load testing the SessionStatus is null)
+
+### Changes in tests
+- Changed document number in tests
+- Added a flag (SmartIdIntegrationTest.TEST_AGAINST_SMART_ID_DEMO) to switch off tests that make requests to Smart-ID demo env.
 
 ## [2.1.2] - 2021-11-03